add interface for arbitrary extra metadata in ES
Akin to the `Extra` fields on structures like `FederationEntityMetadata`
or `OpenIDProviderMetadata`, this commit adds `Extra` to `pkg.Metadata`,
allowing users of `go-oidfed` to insert arbitrary extra metadata
dictionaries into an entity statement without `go-oidfed` needing to
implement the particular entity type. We add a corresponding field to
struct `MetadataPolicies` so that policies can be provided and evaluated
on extra metadata. And we add a method on `Metadata` to look up metadata
by entity type and deserialize it into a particular structure.